Refactor structure of value objects.
